+++ NEWS.md
@@ -1,5 +1,6 @@
 # tglkmeans 0.4.0
 
+* Default of `id_column` parameter was changed to `FALSE`. Note that this is a breaking change, and if you want to use an id column, you need to set it explicitly to `TRUE`.
 * Use R random number generator instead of C++11 random number generator. For backwards compatibility, the old random number generator can be used by setting `use_cpp_random` to `TRUE`.
 * Added parallelization using `RcppParallel`.